% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastshap.R
\name{autoplot.fastshap}
\alias{autoplot.fastshap}
\title{Plotting Shapley values}
\usage{
\method{autoplot}{fastshap}(object, type = c("importance", "dependence"),
  feature = NULL, X, ...)
}
\arguments{
\item{object}{An object of class \code{"fastshap"}.}

\item{type}{Character string specifying which type of plot to construct. 
Current options are \code{"importance"} (for Shapley-based variable 
importance plots) and \code{"dependence"} (for Shapley-based dependence 
plots).}

\item{feature}{Character string specifying which feature to use when 
\code{type = "dependence"}. If \code{NULL} (default) the first feature will
be used to construct the plot.}

\item{X}{A matrix-like R object (e.g., a data frame or matrix) containing 
ONLY the feature columns from the training data.}

\item{...}{Additional optional arguemnts. (Currently ignored.)}
}
\description{
Construct Shap-based importance plots or Shap-based dependence plots.
}
