% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeswarm.R
\name{beeswarm_plot}
\alias{beeswarm_plot}
\title{Internal function to create the beeswarm plot}
\usage{
beeswarm_plot(plt_dat)
}
\arguments{
\item{plt_dat}{A tibble created with internal function
\code{beeswarm_prepare}.}
}
\value{
A ggplot2 object containing the beeswarm plot.
}
\description{
Creates a ggplot2-based beeswarm plot
}
