% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastshap-package.R
\docType{package}
\name{fastshap-package}
\alias{fastshap}
\alias{fastshap-package}
\title{fastshap: Fast Approximate Shapley Values}
\description{
Computes fast (relative to other implementations) approximate Shapley values for any supervised learning model. Shapley values help to explain the predictions from any black box model using ideas from game theory; see Strumbel and Kononenko (2014) \doi{10.1007/s10115-013-0679-x} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bgreenwell/fastshap}
  \item Report bugs at \url{https://github.com/bgreenwell/fastshap/issues}
}

}
\author{
\strong{Maintainer}: Brandon Greenwell \email{greenwell.brandon@gmail.com} (\href{https://orcid.org/0000-0002-8120-0084}{ORCID})

}
\keyword{internal}
