% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastshap.R
\name{fastshap}
\alias{fastshap}
\title{Fast approximate Shapley values}
\usage{
fastshap(object, feature_names = NULL, X, pred_wrapper, nsim = 1, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{feature_names}{Character string giving the names of the predictor
variables (i.e., features) of interest. If \code{NULL} (default) they will be
taken from the column names of \code{X}.}

\item{X}{A matrix-like R object (e.g., a data frame or matrix) containing 
ONLY the feature columns from the training data.}

\item{pred_wrapper}{Prediction function that requires two arguments,
\code{object} and \code{newdata}. The output of this function should be 
determined according to:

\describe{
  \item{Regression}{A numeric vector of predicted outcomes.}
  \item{Binary classification}{A vector of predicted class probabilities
  for the reference class.}
  \item{Multiclass classification}{A vector of predicted class probabilities
  for the reference class.}
}}

\item{nsim}{The number of Monte Carlo repititions to use for estimating each 
Shapley value. Default is 1.}

\item{...}{Additional optional arguments to be passed onto
\code{\link[plyr]{laply}}.}
}
\description{
Compute fast (approximate) Shapley values for a set of features.
}
