% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceplot.R
\name{force_plot}
\alias{force_plot}
\alias{force_plot.explain}
\title{Additive force plot}
\usage{
force_plot(object, ...)

\method{force_plot}{explain}(
  object,
  prediction = NULL,
  baseline = NULL,
  feature_values = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"explain"}.}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{prediction}{Numeric vector containing the corresponding prediction for 
each row in \code{object}.}

\item{baseline}{Numeric giving the average prediction across all the training
observations.}

\item{feature_values}{A matrix-like R object (e.g., a data frame or matrix) 
containing the corresponding feature values for the explanations in 
\code{object}.}
}
\description{
Visualize Shapley values with additive force style layouts from the Python
\href{https://github.com/slundberg/shap}{shap} package.
}
\examples{
#
# A projection pursuit regression (PPR) example
#

# Load the sample data; see ?datasets::mtcars for details
data(mtcars)

# Fit a projection pursuit regression model
mtcars.ppr <- ppr(mpg ~ ., data = mtcars, nterms = 1)

# Compute approximate Shapley values using 10 Monte Carlo simulations
set.seed(101)  # for reproducibility
shap <- explain(mtcars.ppr, X = subset(mtcars, select = -mpg), nsim = 10, 
                pred_wrapper = predict)

# Visualize first explanation
preds <- predict(mtcars.ppr, newdata = mtcars)
x <- subset(mtcars, select = -mpg)[1L, ]  # take first row of feature values
force_plot(shap[1L, ], prediction = preds[1L], baseline = mean(preds),
           feature_values = x)
}
