% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grid.arrange}
\alias{grid.arrange}
\title{Arrange multiple grobs on a page}
\usage{
grid.arrange(..., newpage = TRUE)
}
\value{
A \code{"gtable"} object; see 
\code{\link[gridExtra]{gridExtra-package}} for details.
}
\description{
See \code{\link[gridExtra]{grid.arrange}} for more details.
}
\keyword{internal}
